Add pkg/dnsutil.RRsets function
This function returns the rrset from a slice of rrs. Lifted from the
dnssec plugin (which is now the only user).

This is *also* needed if we're doing per RRset caching which prompted
this PR.

Pros for merging: this function is now tested
Cons: might be less useful than I think

Signed-off-by: Miek Gieben <miek@miek.nl>